requirements.txt
# Core
anthropic>=0.18.0
typer[all]>=0.9.0
rich>=13.0.0
python-dotenv>=1.0.0
